requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
matplotlib==3.9.0
jupyterlab==4.2.5
ipykernel==6.29.5
pyyaml==6.0.2
mlflow==2.14.1
joblib==1.4.2
flask==3.0.3
scipy==1.13.1